refactor: Move `formatFields` utility and its tests to a dedicated file and update imports.
